+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_help.py
@@ -807,6 +807,11 @@ helps['vm image show'] = """
             az vm image show -l westus -f CentOS -p OpenLogic --s 7.3 --version {latest}
 """
 
+helps['vm image accept-term'] = """
+    type: command
+    short-summary: Accept Azure Marketplace term so that the image can be used to create VMs
+"""
+
 helps['vm nic'] = """
     type: group
     short-summary: Manage network interfaces. See also `az network nic`.